fix access issue

+++ commands/tiktok.js
@@ -1,12 +1,12 @@
+const fs = require("fs");
 const needle = require("needle");
-
 const bot = require("../bot");
 const redisClient = require('../redisClient');
 const fetch = (...args) => import('node-fetch').then(({default: fetch}) => fetch(...args));
 const puppeteer = require("puppeteer");
 
 const REGEXP_LINK = /^(https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|www\.[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9]+\.[^\s]{2,}|www\.[a-zA-Z0-9]+\.[^\s]{2,})$/i;
-
+const MAX_TRY = 5;
 bot.hears(REGEXP_LINK, async (ctx) => {
     await sendTikTokVideo(ctx);
 })
@@ -16,27 +16,39 @@ bot.command('tiktok', async ctx => {
 })
 
 const sendTikTokVideo = async (ctx) => {
+    let currentTry = 0;
+    let filePath = null;
     console.log(`-> from ${ctx.message.from.first_name || ''} ${ctx.message.from.last_name || ''}: ${ctx.message.text}`)
     const tikTokLink = ctx.message.text.replace('/tiktok', '').replaceAll(' ', '');
     if (tikTokLink.toLowerCase().includes('tiktok.com') && REGEXP_LINK.test(tikTokLink)) {
         try {
             bot.telegram.sendChatAction(ctx.chat.id, 'record_video');
-            const videoLink = await getTikTokVideoLinkWithNeedle(tikTokLink);
-            console.log('download link', videoLink)
-            const video = await fetch(videoLink.video.url);
-            const buffer = await video.buffer();
+            let videoInfo = null;
+            do {
+                videoInfo = await getTikTokVideoLinkWithNeedle(tikTokLink);
+                filePath = `tiktok-${videoInfo.id}.mp4`;
+                console.log('download link', videoInfo)
+                await downloadFile(videoInfo.video.url, filePath);
+                currentTry++;
+                console.log('file size',getFilesizeInMegabytes(filePath))
+            } while (getFilesizeInMegabytes(filePath) < 0.1 && currentTry < MAX_TRY)
 
-            //send video to chat
-            bot.telegram.sendChatAction(ctx.chat.id, 'upload_video');
-            await ctx.replyWithVideo({source: buffer}, {
-                caption:
-                    `-> ${ctx.message.from.first_name || ''} ${ctx.message.from.last_name || ''}\n${videoLink.video.description}`
-            })
+            if (getFilesizeInMegabytes(filePath) < 0.1){
+                throw new Error(`Чот не получилось))`);
+            } else {
+                //send video to chat
+                bot.telegram.sendChatAction(ctx.chat.id, 'upload_video');
+                await ctx.replyWithVideo({source: filePath}, {
+                    caption:
+                        `-> ${ctx.message.from.first_name || ''} ${ctx.message.from.last_name || ''}\n${videoInfo.video.description}`
+                })
 
-            //delete video url
-            ctx.deleteMessage(ctx.message.id);
-            //save cookies
-            await redisClient.set("lastSuccessCookies", JSON.stringify(videoLink.cookies));
+                //delete video url
+                ctx.deleteMessage(ctx.message.id);
+                //save cookies
+                await redisClient.set("lastSuccessCookies", JSON.stringify(videoInfo.cookies));
+            }
+            fs.unlinkSync(filePath);
         } catch (error) {
             console.log(`<- to ${ctx.message.from.first_name || ''} ${ctx.message.from.last_name || ''}: TikTok Video Failed - ${error}`)
             ctx.reply(`ERROR TikTok: ${error}`)
@@ -61,12 +73,28 @@ const getTikTokVideoLinkWithNeedle = async (link) => {
     const videoURL = videoObject.ItemModule[id].video.downloadAddr.trim();
     const videoDescription = videoObject.ItemModule[id].desc;
 
-    console.log("VideoObject", videoObject.ItemModule[id]);
     return {
+        id,
         cookies,
         video: {
             url: videoURL,
             description: videoDescription
         }
     }
+}
+
+const downloadFile = async (url, path) => {
+    const res = await fetch(url);
+    const fileStream = fs.createWriteStream(path);
+    await new Promise((resolve, reject) => {
+        res.body.pipe(fileStream);
+        res.body.on("error", reject);
+        fileStream.on("finish", resolve);
+    });
+}
+
+const getFilesizeInMegabytes = (filename) => {
+    const stats = fs.statSync(filename);
+    const fileSizeInBytes = stats.size;
+    return fileSizeInBytes / (1024*1024);
 }
